add an option to control progress reporting
